authorize-gate-islemler

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,7 @@ use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -34,6 +35,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_post');
         $categories = Category::all();
         return view('create', compact('categories'));
     }
@@ -43,6 +45,7 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create_post');
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', 'max:255'],
@@ -75,6 +78,7 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit_post');
         $post = Post::findOrFail($id);
         $categories = Category::all();
         return view('edit', compact('post', 'categories'));
@@ -85,7 +89,7 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
-
+        $this->authorize('edit_post');
         $request->validate([
             'title' => ['required', 'max:255'],
             'category_id' => ['required', 'integer'],
@@ -115,20 +119,24 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_post');
         $post = Post::findOrFail($id);
         $post->delete();
         return redirect()->route('posts.index');
     }
     public function trashed() {
+        $this->authorize('delete_post');
         $posts = Post::onlyTrashed()->get();
         return view('trashed',compact('posts'));
     }
     public function restore($id){
+        $this->authorize('delete_post');
         $post = Post::onlyTrashed()->findOrFail($id);
         $post->restore();
         return redirect()->back();
     }
     public function forceDelete($id) {
+        $this->authorize('delete_post');
         $post = Post::onlyTrashed()->findOrFail($id);
         File::delete(public_path($post->image));
         $post->forceDelete();
